+++ packages/crypto-signature-schnorr/source/service-provider.ts
@@ -1,0 +1,21 @@
+import { Identifiers } from "@mainsail/constants";
+import { injectable, Selectors } from "@mainsail/container";
+import { Providers } from "@mainsail/kernel";
+
+import { Signature } from "./signature.js";
+
+@injectable()
+export class ServiceProvider extends Providers.ServiceProvider {
+	public async register(): Promise<void> {
+		this.app
+			.bind(Identifiers.Cryptography.Signature.Size)
+			.toConstantValue(64)
+			.when(Selectors.anyAncestorOrTargetTagged("type", "wallet"));
+
+		this.app
+			.bind(Identifiers.Cryptography.Signature.Instance)
+			.to(Signature)
+			.inSingletonScope()
+			.when(Selectors.anyAncestorOrTargetTagged("type", "wallet"));
+	}
+}
